Initial version of loading auth before moving higher up the tree
